refactor(indexing): split project workflow planning
Part of #1108.
